refactor: entities and value objects
