+++ internal/domain/repository/factory.go
@@ -3,4 +3,5 @@ package repository
 type Factory interface {
 	NewUserRepository() UserRepository
 	NewAuthenticationRepository() AuthenticationRepository
+	NewStoreRepository() StoreRepository
 }
