+++ internal/domain/repository/user.go
@@ -12,6 +12,7 @@ type UserRepository interface {
 	FindByEmail(ctx context.Context, email string) (entity.User, error)
 	// GetList(ctx context.Context, items int, page int) ([]*User, error)
 	Update(ctx context.Context, user entity.User) error
+	AddRoles(ctx context.Context, userID string, roles []entity.UserRole) error
 	ResetPasswordByEmail(ctx context.Context, id string) error
 	Delete(ctx context.Context, id string) error
 }
